+++ frontend/src/components/constants/visualizationCategories.js
@@ -1,0 +1,79 @@
+export const CLUSTER_CATEGORY_MAPPINGS = {
+	learning: "Learning Focused",
+	educational: "Educational",
+	workshop: "Workshop",
+	seminar: "Seminar",
+	training: "Training",
+	lecture: "Lecture",
+	presentation: "Presentation",
+	course: "Course",
+	tutorial: "Tutorial",
+	content: "Valuable Content",
+
+	networking: "Networking",
+	social: "Social",
+	meetup: "Meetup",
+	community: "Community",
+	connection: "Connection",
+	relationship: "Relationship",
+	friendship: "Friendship",
+
+	career: "Career Development",
+	professional: "Professional",
+	job: "Job",
+	employment: "Employment",
+	industry: "Industry",
+	business: "Business",
+	entrepreneur: "Entrepreneurial",
+
+	technical: "Technical",
+	technology: "Technology",
+	coding: "Coding",
+	programming: "Programming",
+	development: "Development",
+	engineering: "Engineering",
+	software: "Software",
+	data: "Data",
+	ai: "AI/ML",
+	machine: "Machine Learning",
+
+	fun: "Fun",
+	entertainment: "Entertainment",
+	game: "Gaming",
+	activity: "Activity",
+	recreation: "Recreation",
+	leisure: "Leisure",
+
+	food: "Food & Dining",
+	lunch: "Lunch",
+	dinner: "Dinner",
+	breakfast: "Breakfast",
+	refreshment: "Refreshments",
+	snack: "Snacks",
+
+	campus: "Campus",
+	university: "University",
+	library: "Library",
+	lab: "Laboratory",
+	office: "Office",
+	conference: "Conference",
+	meeting: "Meeting",
+	room: "Room",
+	hall: "Hall",
+	center: "Center",
+	building: "Building",
+};
+
+export const CLUSTER_SIZE_THRESHOLDS = {
+	LARGE: 8,
+	VERY_LARGE: 10,
+	MEDIUM: 4,
+};
+
+export const FALLBACK_NAMES = {
+	EXTENDED_LEARNING: "Extended Learning Events",
+	HIGH_IMPACT: "High-Impact Events",
+	DIVERSE_MIX: "Diverse Event Mix",
+	SPECIALIZED: "Specialized Events",
+	NICHE: "Niche Events",
+};
